refactor(ChessGame): 기물 이동 검증 클래스 분리
- ChessGame 내에서 검증했던 기물의 이동을 따로 클래스로 분리하여 구현
- 패키지 분리
