+++ src/main/java/softeer2nd/chess/game/Board.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess;
+package softeer2nd.chess.game;
 
 import softeer2nd.chess.pieces.Piece;
