+++ src/main/java/softeer2nd/chess/game/GameStatus.java
@@ -1,4 +1,4 @@
-package softeer2nd.chess;
+package softeer2nd.chess.game;
 
 import softeer2nd.chess.pieces.Piece.Color;
