+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import softeer2nd.chess.exceptions.BoardIndexOutOfRangeException;
+import softeer2nd.chess.game.Board;
+import softeer2nd.chess.game.Position;
 import softeer2nd.chess.pieces.Piece;
 import softeer2nd.chess.pieces.Piece.Type;
