+++ src/test/java/softeer2nd/chess/ChessViewTest.java
@@ -3,6 +3,8 @@ package softeer2nd.chess;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import softeer2nd.chess.game.Board;
+import softeer2nd.chess.game.ChessView;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static softeer2nd.chess.utils.StringUtils.appendNewLine;
