+++ src/test/java/softeer2nd/chess/pieces/KingTest.java
@@ -3,9 +3,9 @@ package softeer2nd.chess.pieces;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.chess.Board;
-import softeer2nd.chess.ChessGame;
-import softeer2nd.chess.Position;
+import softeer2nd.chess.game.Board;
+import softeer2nd.chess.game.ChessGame;
+import softeer2nd.chess.game.Position;
 import softeer2nd.chess.exceptions.IllegalMoveException;
 
 import java.util.List;
